requirements.txt
google-adk
google-generativeai
python-dotenv
sqlalchemy
aiosqlite
greenlet